+++ R/env.R
@@ -1,2 +1,0 @@
-pkg.env <- new.env(parent = emptyenv())
-pkg.env$praat.path <- "./"